lol hey again
new index!!! now all in three columns and one row!!! epic!!!
the padding is off but i'll try my best to fix it tomorrow
it's..... just turned to 1:10 am lol i was waiting for it for (just) a bit to turn to it there so that this'd be easier to write/mention/talk about. anyways.
so yeah!! i'll make this look slightly nicer tomorrow too, i guess. it's Mostly fine already just copying and pasting it. made some tiny little changes to it too already, btw.