more padding again lol
i think i actually do need it, some more??? in just one tab mode (i think??) it showed some of the footer. to be fair/to be clear, this Still shows some of the footer but at least it's more subtle now and i think it looks nice like this.