oops some more emojis to fix lol
news archive one should be different because i picked the full newspaper not the rolled up one
art Should be the same?

oh lemme change it to just "art" as well, Hopefully that's not sus but like it's all Mine right so it's redundant. let's hope it doesn't look Too unbalanced!