EXPLORE THE SITE + RENAMED AND NEW PAGES + HAPPY 700TH COMMIT AWESOME
also no more javascript on the site as time thing is gone, not so useful and also a bit less than a second behind! woohoo!!
some of these pages are yet to be created or renamed. well... all of them i mean other than the current ones.
navigation is gonna get renamed to sitemap if i do something with it after this yeah, wow!!
about me is up next, baby!!!!!! yipeee!
also adjusted speed and extent of floating animation (float!)
stuff seems off center but idk
also i added a link to discord account cause it wasn't glowing or changing the mouse cursor like every single other link and it was bugging me
ALSO and the final thing for this huge commit, swapped the link and the span in the socials box so that the WHOLE box is the link not just the image and the text, so now it is way smoother moving between one account box and another! it seems it's MUCH better, i can go with my mouse from one account to the next and there is NO gap between them!!