Create 2022-02-13-one-year-anniversary.html
first of the update post pages baby!!!!!!!
these titles are not auto-generated i gotta write it out this technically myself lol
idk Exactly how i feel about this format but eh it is alright so far
different back button link colour to news archive as well! let's see if i change it, not so sure on it yet.
changed it (the link colour) a bit and you know what, i like this now!